Divide

+++ IMC.Testing.HelloWorld/Functions.cs
@@ -4,9 +4,11 @@ namespace IMC.Testing.HelloWorld
 {
     public class Functions
     {
-        public double Divide(double a, double b)
+        public double Divide(int a, int b)
         {
             return a / b;
         }
+
+
     }
 }
